chore: Rename near_types to types. Remove unused deps
