+++ clients/implementations/near/src/lib.rs
@@ -1,3 +1,3 @@
 pub mod client;
-pub mod near_types;
+pub mod types;
 pub mod wiring_layer;
